% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gen_general.R
\name{blaP}
\alias{blaP}
\title{Print nothing}
\usage{
blaP(disp = "S")
}
\arguments{
\item{disp}{character length one it is a parameter}
}
\value{
print the string
}
\description{
hhh
}
\details{
sss
}

