% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_import.R
\name{dataImport}
\alias{dataImport}
\title{Import Data}
\usage{
dataImport(doPar = TRUE, scanIdCol = "def", confIdCol = "def",
  NrAdjust = 1)
}
\arguments{
\item{doPar}{Logical. If datafile import should be done in parallel.}

\item{scanIdCol}{The standard colum name for the scanner Id.}

\item{confIdCol}{The standard colum name for the configuration Id.}

\item{NrAdjust}{Defaults to 1 XXX}
}
\value{
A (big) list.
}
\description{
Imports data from the folder-structure in the 'rawdata' folder
}
\details{
XXX Here the details of how the folder should be named, with 
separators etc.
}

