% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_import.R
\name{mDataStr}
\alias{mDataStr}
\title{add variables to ET data}
\usage{
mDataStr(rawData, day = 1, nameChange = FALSE)
}
\arguments{
\item{rawData}{ET raw data imported from the txt file}

\item{day}{optional argument useful if the experiment was performed in different days}

\item{nameChange}{optional argument if the name of the groups has to be renamed}
}
\value{
a table with additional columns
}
\description{
Add variables
}
\details{
XXX Here the details of how the folder should be named, with 
separators etc.
}
