% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TellSpec.R
\docType{package}
\name{TellSpec}
\alias{TellSpec}
\alias{TellSpec-package}
\title{TellSpec}
\description{
Doing things that need to be done - washing for example
}
\details{
Better explanation here
}
\author{
we
}

